# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KOSMOS API",
    "description": "KOSMOS AI Governance Platform - Multi-agent orchestration with Pentarchy voting.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/mcp/servers": {
      "get": {
        "tags": [
          "mcp"
        ],
        "summary": "List Servers",
        "description": "List all registered MCP servers.",
        "operationId": "list_servers_api_v1_mcp_servers_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ServerStatus"
                  },
                  "type": "array",
                  "title": "Response List Servers Api V1 Mcp Servers Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "mcp"
        ],
        "summary": "Register Server",
        "description": "Register a new MCP server.",
        "operationId": "register_server_api_v1_mcp_servers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServerRegistration"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mcp/servers/{server_id}": {
      "delete": {
        "tags": [
          "mcp"
        ],
        "summary": "Unregister Server",
        "description": "Unregister an MCP server.",
        "operationId": "unregister_server_api_v1_mcp_servers__server_id__delete",
        "parameters": [
          {
            "name": "server_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Server Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/mcp/tools": {
      "get": {
        "tags": [
          "mcp"
        ],
        "summary": "List Tools",
        "description": "List available tools, optionally filtered by server or query.",
        "operationId": "list_tools_api_v1_mcp_tools_get",
        "parameters": [
          {
            "name": "server_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Server Id"
            }
          },
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Query"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ToolInfo"
                  },
                  "title": "Response List Tools Api V1 Mcp Tools Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mcp/tools/{server_id}/{tool_name}": {
      "get": {
        "tags": [
          "mcp"
        ],
        "summary": "Get Tool",
        "description": "Get details of a specific tool.",
        "operationId": "get_tool_api_v1_mcp_tools__server_id___tool_name__get",
        "parameters": [
          {
            "name": "server_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Server Id"
            }
          },
          {
            "name": "tool_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tool Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mcp/tools/{server_id}/{tool_name}/invoke": {
      "post": {
        "tags": [
          "mcp"
        ],
        "summary": "Invoke Tool",
        "description": "Invoke a tool on a specific server.",
        "operationId": "invoke_tool_api_v1_mcp_tools__server_id___tool_name__invoke_post",
        "parameters": [
          {
            "name": "server_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Server Id"
            }
          },
          {
            "name": "tool_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tool Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolInvocation"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ServerRegistration": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Id"
          },
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "url": {
            "type": "string",
            "title": "Url",
            "default": ""
          },
          "transport": {
            "type": "string",
            "pattern": "^(http|stdio|websocket|mock)$",
            "title": "Transport",
            "default": "mock"
          },
          "command": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Command"
          },
          "args": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Args"
          },
          "env": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Env"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "ServerRegistration",
        "description": "Request to register a new MCP server."
      },
      "ServerStatus": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "tools_count": {
            "type": "integer",
            "title": "Tools Count"
          },
          "last_health_check": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Health Check"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "status",
          "tools_count",
          "last_health_check",
          "error"
        ],
        "title": "ServerStatus",
        "description": "MCP server status response."
      },
...
      "ToolInfo": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "input_schema": {
            "additionalProperties": true,
            "type": "object",
            "title": "Input Schema"
          },
          "server_id": {
            "type": "string",
            "title": "Server Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "description",
          "input_schema",
          "server_id"
        ],
        "title": "ToolInfo",
        "description": "Tool information."
      },
      "ToolInvocation": {
        "properties": {
          "arguments": {
            "additionalProperties": true,
            "type": "object",
            "title": "Arguments"
          }
        },
        "type": "object",
        "title": "ToolInvocation",
        "description": "Request to invoke a tool."
      },
      "ToolResult": {
        "properties": {
          "server_id": {
            "type": "string",
            "title": "Server Id"
          },
          "tool_name": {
            "type": "string",
            "title": "Tool Name"
          },
          "result": {
            "additionalProperties": true,
            "type": "object",
            "title": "Result"
          },
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "server_id",
          "tool_name",
          "result",
          "success"
        ],
        "title": "ToolResult",
        "description": "Tool invocation result."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}